backend/requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.3
flask-cors==6.0.2
gunicorn==23.0.0
ImageHash==4.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.3
pillow==12.1.1
PyWavelets==1.9.0
rapidfuzz==3.8.1
scipy==1.17.1
Werkzeug==3.1.7
